Merge commit '4c9f9c818223cc5ad0db63553796a54f29e959d1'

* commit '4c9f9c818223cc5ad0db63553796a54f29e959d1':
  remove floats from libcamera2

+++ libcamera2/QualcommCameraHardware.h
@@ -125,12 +125,12 @@ private:
     int mPreviewWidth;
     int mRawHeight;
     int mRawWidth;
-	unsigned int frame_size;
-	int mBrightness;
-	float mZoomValuePrev;
-    float mZoomValueCurr;
-	bool mZoomInitialised;
-	bool mCameraRunning;
+    unsigned int frame_size;
+    int mBrightness;
+    int mZoomValuePrev;
+    int mZoomValueCurr;
+    bool mZoomInitialised;
+    bool mCameraRunning;
     bool mPreviewInitialized;
     
     // This class represents a heap which maintains several contiguous
